+++ packages/config/src/configs/moonbaseAlpha.ts
@@ -1,12 +1,25 @@
 import { BalanceBuilder, ContractBuilder } from '@moonbeam-network/xcm-builder';
-import { dev, eq, eqd, lit, paring, tt1, tur, unit } from '../assets';
+import {
+  dev,
+  eq,
+  eqd,
+  lit,
+  nodl,
+  paring,
+  soon,
+  tt1,
+  tur,
+  unit,
+} from '../assets';
 import {
   alphanetRelay,
   darwiniaPangoro,
   equilibriumAlphanet,
   litentryAlphanet,
   moonbaseAlpha,
+  nodleEden,
   statemineAlphanet,
+  subsocialAlphanet,
   turingAlphanet,
 } from '../chains';
 import { AssetConfig } from '../types/AssetConfig';
@@ -82,6 +95,20 @@ export const moonbaseAlphaConfig = new ChainConfig({
       },
     }),
     new AssetConfig({
+      asset: nodl,
+      balance: BalanceBuilder().assets().account(),
+      contract: ContractBuilder().Xtokens().transfer(),
+      destination: nodleEden,
+      destinationFee: {
+        amount: 0.02,
+        asset: nodl,
+      },
+      fee: {
+        asset: dev,
+        balance: BalanceBuilder().system().account(),
+      },
+    }),
+    new AssetConfig({
       asset: paring,
       balance: BalanceBuilder().assets().account(),
       contract: ContractBuilder().Xtokens().transfer(),
@@ -89,6 +116,20 @@ export const moonbaseAlphaConfig = new ChainConfig({
       destinationFee: {
         amount: 4,
         asset: paring,
+      },
+      fee: {
+        asset: dev,
+        balance: BalanceBuilder().system().account(),
+      },
+    }),
+    new AssetConfig({
+      asset: soon,
+      balance: BalanceBuilder().assets().account(),
+      contract: ContractBuilder().Xtokens().transfer(),
+      destination: subsocialAlphanet,
+      destinationFee: {
+        amount: 0.4,
+        asset: soon,
       },
       fee: {
         asset: dev,